rename items folder to sections
