+++ app/pages/sections/description/manifest.json
