+++ app/routes.test.js
@@ -5,7 +5,7 @@ import { routes } from './routes';
 import { baseUrl } from './config';
 import { getCsrfTokenFromGet } from './test-utils/helper';
 import * as dashboardController from './pages/dashboard/controller';
-import * as descriptionController from './pages/items/description/controller';
+import * as descriptionController from './pages/sections/description/controller';
 
 jest.mock('./logger');
